Fix arithmetic tester calc scripts.
Add ?e+2047 to the ones that didn't have it (this caused a spurious
test failure for me) and place it before ?b16 for the ones that already
did.  Doing ?e+ before ?b is not really important; I just feel a little
better doing it while the base is still set to decimal, even though I
know perfectly well the argument is parsed as decimal regardless of the
base setting.